==> experiment-1/analysis/03_descriptive_stats_session1.Rmd <==
---
title: 'Descpriptive statistics of Session 1'
author: "Maarten van der Velde"
date: "Last updated: `r Sys.Date()`"
output:
  html_notebook:
    smart: no
    toc: yes
  html_document:
    df_print: paged
    toc: yes
editor_options: 
  chunk_output_type: inline
---

# Setup
```{r}
library(dplyr)
library(forcats)
library(ggplot2)
library(jsonlite)
library(purrr)
library(scales)
library(tidyr)
library(fst)

theme_set(theme_light(base_size = 14))
```


---

# Session 1

## Lab data

```{r}
lab_1_rl <- read.fst(file.path("..", "data", "processed", "lab", "session1", "session1_rl_lab.fst"))
lab_1_rl_final_alpha <- read.fst(file.path("..", "data", "processed", "lab", "session1", "session1_final_alpha_lab.fst"))
lab_1_typing <- read.fst(file.path("..", "data", "processed", "lab", "session1", "session1_typing_lab.fst"))
```



### Participants

There were `r length(unique(lab_1_rl$subject))` participants in the lab.

### Session stats
```{r}
session_stats <- lab_1_rl %>%
  group_by(subject) %>%
  summarise(n_trials = n(),
  session_duration = (max(start_time) - min(start_time) + if_else(rt[which.max(start_time)] == Inf, median(rt), rt[which.max(start_time)])) / 60000,
  perc_correct = mean(correct),
  n_items = length(unique(fact_id)))
```


The boxplots below show the distributions of some performance statistics calculated for each participant's session.

The duration of the session was fixed at 10 minutes (participants were allowed to finish their final trial), which is confirmed by the first plot.
The reconstructed duration is the difference between the start time of the first trial and the start time of the last trial, plus the response time on the last trial (an infinite RT on the last trial was replaced by the participant's median RT).

The number of trials varied substantially between participants but looks to be normally distributed (second plot).

The number of unique facts that a participant encountered also varied strongly (third plot).

The mean accuracy within a session varied from about 30% to about 95%.


```{r}
par(mfrow=c(2,2))
boxplot(session_stats$session_duration, horizontal = TRUE, main = "Session duration (min)")
boxplot(session_stats$n_trials, horizontal = TRUE, main = "Number of trials")
boxplot(session_stats$n_items, horizontal = TRUE, main = "Number of unique facts")
boxplot(session_stats$perc_correct, horizontal = TRUE, main = "Proportion correct")
```


### Fact alpha

We will use this dataset to derive fact alpha predictions

#### How many participants saw each fact?

```{r}
times_seen <- lab_1_rl %>%
  group_by(fact_id) %>%
  summarise(n = length(unique(subject)))
```

Each fact was seen once or more by at least `r min(times_seen$n)` participants and by at most `r max(times_seen$n)` participants.

However, for a useful alpha estimate we need at least 3 encounters -- only then is some adjustment in alpha possible.

```{r}
times_seen_usable <- lab_1_rl %>%
  group_by(subject, fact_id) %>%
  filter(n() >= 3) %>%
  group_by(fact_id) %>%
  summarise(n = length(unique(subject)))
```

Each fact was seen three times or more by at least `r min(times_seen_usable$n)` participants and by at most `r max(times_seen_usable$n)` participants.

#### Final alpha distribution

```{r}
final_alphas <- lab_1_rl_final_alpha %>%
  filter(reps >= 3)

hist(final_alphas$final_alpha, main = "Final alpha (facts with >= 3 repetitions)")
```


#### Final alpha per fact

The red dashed line shows the default starting alpha of 0.3.
```{r}
ggplot(final_alphas, aes(x = reorder(as.factor(fact_id), final_alpha, FUN = median), y = final_alpha, group = as.factor(fact_id))) +
  geom_hline(yintercept = 0.3, colour = "red", lty = 2) +
  geom_boxplot(width = 0.5) +
  geom_point(alpha = 0.5, size = 0.5) %>%
  labs(x = "Fact ID",
       y = "Final rate of forgetting")
```


### Typing performance

Participants also completed a 6-item typing test at the start of the session, to detect any difficulties with typing.

```{r}
lab_typing <- lab_1_typing %>%
  select(subject, keypresses, answer, correct, backspace_used, backspaced_first_letter)
```

Lab participants were virtually flawless in the accuracy of their final response, as the boxplot below shows. 

```{r}
lab_typing_accuracy <- lab_typing %>%
  group_by(subject) %>%
  summarise(accuracy = sum(correct) / n()) %>%
  pull(accuracy)
lab_typing_accuracy %>%
  boxplot(ylim = c(0, 1), horizontal = TRUE, xlab = "Accuracy")
```

```{r}
table(lab_typing_accuracy)
```


The plot below shows the key press times per subject and per test item.
The items are plotted in the order they were shown (top left to bottom right).
Key press time is the time difference between consecutive key presses, except in the case of the first key press, when it is the time since the appearance of the stimulus.

The plots suggest that the first key press on the first item is generally slower than the first key press on subsequent items, although it is clear that participants are suddenly slower on *zipper*, most likely due to the uncommon first character.

```{r}
lab_typing_by_key <- lab_typing %>%
  mutate(y = map(keypresses, ~fromJSON(as.character(.x)))) %>%
  unnest() %>%
  select(-keypresses) %>%
  group_by(subject, answer) %>%
  mutate(key_num = 1:n(),
         key_time = rt - lag(rt),
         key_time = ifelse(is.na(key_time), rt, key_time)) %>%
  ungroup()

lab_typing_by_key_correct <- lab_typing_by_key %>%
  filter(correct, !backspace_used, !backspaced_first_letter)

lab_typing_by_key_correct %>%
  ggplot(aes(x = as.factor(key_num), y = key_time, group = subject, colour = subject)) +
  geom_line(alpha = 0.5) +
  scale_y_log10() +
  facet_wrap(~ answer) +
  guides(colour = FALSE) +
  labs(title = "Key press times by subject on the six typing test items (Lab)",
       subtitle = "Only correct responses. Key press time is the time difference between consecutive key presses.",
       x = "Character #",
       y = "Key press time (ms) (log-scale)")
  
```


Based on these key press times we can calculate a participant's typing speed.
For each participant, we calculate the *median* key press time across all their key presses in the typing test (including those that are incorrect) and derive a raw words-per-minute value based on a word-length of 5 letters.


```{r}
lab_typing_speed <- lab_typing_by_key %>%
  group_by(subject) %>%
  summarise(key_time = median(key_time),
            wpm = 60000 / (key_time * 5))

lab_typing_speed$wpm %>%
  boxplot(horizontal = TRUE, main = "Typing speed of lab participants", xlab = "Raw words per minute")
```

Typing speed data is not easy to find, but this source shows that the mode of the distribution lies around 30 WPM:
https://www.livechatinc.com/typing-speed-test/#/global-scores


## MTurk

```{r}
mturk_1_rl <- read.fst(file.path("..", "data", "processed", "mturk", "session1", "session1_rl_mturk.fst"))
mturk_1_rl_final_alpha <- read.fst(file.path("..", "data", "processed", "mturk", "session1", "session1_final_alpha_mturk.fst"))
mturk_1_typing <- read.fst(file.path("..", "data", "processed", "mturk", "session1", "session1_typing_mturk.fst"))
mturk_1_questionnaire <- read.fst(file.path("..", "data", "processed", "mturk", "session1", "session1_questionnaire_mturk.fst"))
```


### Participants

There were `r length(unique(mturk_1_rl$subject))` participants in the MTurk sample.
These participants took part in one of three batches run on the MTurk website between 4 March 2019 and 19 March 2019.


### Demographics

The MTurk participants were recruited with the requirement that they were located in the US, had a US High School graduate degree, and were born in the 1992-1999 bracket (age 18-25). They also completed a two-item questionnaire about their current level of education and occupation.
The distributions of responses are shown below.

```{r, fig.width=10}
mturk_1_questionnaire %>%
  count(degree) %>%
  mutate(perc = n / sum(n)) %>%
  mutate(degree = fct_relevel(degree, 
                              "Master's degree (e.g., MA, MS, MEd)",
                              "Bachelor's degree (e.g., BA, BS)",
                              "Associate degree (e.g., AA, AS)",
                              "High school degree or equivalent (e.g. GED)"
                              )) %>%
  ggplot(aes(degree, perc)) +
  geom_col() +
  coord_flip() +
  scale_y_continuous(labels = percent_format()) +
  labs(title = "Highest degree or level of school completed?", x = "", y = "Frequency")
```

```{r, fig.width = 10}
mturk_1_questionnaire %>%
  count(occupation) %>%
  mutate(perc = n / sum(n)) %>%
  mutate(occupation = fct_relevel(occupation, 
                              "Other",
                              "Student",
                              "Employed or self-employed (full time or part time)"
                              )) %>%
  ggplot(aes(occupation, perc)) +
  geom_col() +
  coord_flip() +
  scale_y_continuous(labels = percent_format()) +
  labs(title = "Current occupational status?", x = "", y = "Frequency")
```



### Session stats
```{r}
session_stats_m <- mturk_1_rl %>%
  group_by(subject) %>%
  summarise(n_trials = n(),
  session_duration = (max(start_time) - min(start_time) + if_else(rt[which.max(start_time)] == Inf, median(rt), rt[which.max(start_time)])) / 60000,
  perc_correct = mean(correct),
  n_items = length(unique(fact_id)))
```


The boxplots below show the distributions of some performance statistics calculated for each participant's session.

The duration of the session was fixed at 10 minutes.
Participants that clicked away during the experiment (n = 7) were not allowed to finish their session and are excluded from the data.
The reconstructed duration is the difference between the start time of the first trial and the start time of the last trial, plus the response time on the last trial (an infinite RT on the last trial was replaced by the participant's median RT).

The number of trials varied substantially between participants but looks to be normally distributed (second plot). MTurk participants seem to have completed slightly more trials than the lab participants.

The number of unique facts that a participant encountered also varied strongly (third plot). Participants on MTurk seem to have managed slightly more unique facts than the lab participants, without sacrificing accuracy, which is very similar in the two groups (fourth plot).


```{r}
par(mfrow=c(2,2))
boxplot(session_stats_m$session_duration, horizontal = TRUE, main = "Session duration (min)")
boxplot(session_stats_m$n_trials, horizontal = TRUE, main = "Number of trials")
boxplot(session_stats_m$n_items, horizontal = TRUE, main = "Number of unique facts")
boxplot(session_stats_m$perc_correct, horizontal = TRUE, main = "Proportion correct")
```



### Fact alpha

We want to use this data set to make fact alpha predictions

#### How many participants saw each fact?

```{r}
times_seen_m <- mturk_1_rl %>%
  group_by(fact_id) %>%
  summarise(n = length(unique(subject)))
```

Each fact was seen once or more by at least `r min(times_seen_m$n)` participants and by at most `r max(times_seen_m$n)` participants.

However, for a useful alpha estimate we need at least 3 encounters -- only then is some adjustment in alpha possible.

```{r}
times_seen_usable_m <- mturk_1_rl %>%
  group_by(subject, fact_id) %>%
  filter(n() >= 3) %>%
  group_by(fact_id) %>%
  summarise(n = length(unique(subject)))
```

Each fact was seen three times or more by at least `r min(times_seen_usable_m$n)` participants and by at most `r max(times_seen_usable_m$n)` participants.

#### Final alpha distribution

```{r}
final_alphas_m <- mturk_1_rl_final_alpha %>%
  filter(reps >= 3)

hist(final_alphas_m$final_alpha, main = "Final alpha (facts with >= 3 repetitions)")
```


#### Final alpha per fact

```{r}
ggplot(final_alphas_m, aes(x = reorder(as.factor(fact_id), final_alpha, FUN = median), y = final_alpha, group = as.factor(fact_id))) +
  geom_hline(yintercept = 0.3, colour = "red", lty = 2) +
  geom_boxplot(width = 0.5) +
  geom_point(alpha = 0.5, size = 0.5) +
  labs(x = "Fact ID",
       y = "Final rate of forgetting")
```



### Typing performance

Participants also completed a 6-item typing test at the start of the session, to detect any difficulties with typing.

```{r}
mturk_typing <- mturk_1_typing %>%
  select(subject, keypresses, answer, correct, backspace_used, backspaced_first_letter)
```

Like the lab participants, MTurk participants were virtually flawless in the accuracy of their final response, as the boxplot below shows. 

```{r}
mturk_typing_accuracy <- mturk_typing %>%
  group_by(subject) %>%
  summarise(accuracy = sum(correct) / n()) %>%
  pull(accuracy)

mturk_typing_accuracy %>%
  boxplot(ylim = c(0, 1), horizontal = TRUE, xlab = "Accuracy")
```

```{r}
table(mturk_typing_accuracy)
```

The plot below shows the key press times per subject and per test item.
The items are plotted in the order they were shown (top left to bottom right).
Key press time is the time difference between consecutive key presses, except in the case of the first key press, when it is the time since the appearance of the stimulus.

The plots look very similar to those of the lab participants.

```{r}
mturk_typing_by_key <- mturk_typing %>%
  mutate(y = map(keypresses, ~fromJSON(as.character(.x)))) %>%
  unnest() %>%
  select(-keypresses) %>%
  group_by(subject, answer) %>%
  mutate(key_num = 1:n(),
         key_time = rt - lag(rt),
         key_time = ifelse(is.na(key_time), rt, key_time)) %>%
  ungroup()

mturk_typing_by_key_correct <- mturk_typing_by_key %>%
  filter(correct, !backspace_used, !backspaced_first_letter)

mturk_typing_by_key_correct %>%
  ggplot(aes(x = as.factor(key_num), y = key_time, group = subject, colour = subject)) +
  geom_line(alpha = 0.5) +
  scale_y_log10() +
  facet_wrap(~ answer) +
  guides(colour = FALSE) +
  labs(title = "Key press times by subject on the six typing test items (MTurk)",
       subtitle = "Only correct responses. Key press time is the time difference between consecutive key presses.",
       x = "Character #",
       y = "Key press time (ms) (log-scale)")
  
```


Based on these key press times we can calculate a participant's typing speed.
For each participant, we calculate the *median* key press time across all their key presses in the typing test (including those that are incorrect) and derive a raw words-per-minute value based on a word-length of 5 letters.

Compared to the lab participants, the MTurk participants are faster typers.
This is not surprising, given that the MTurk participants also completed more trials in the learning session.

```{r}
mturk_typing_speed <- mturk_typing_by_key %>%
  group_by(subject) %>%
  summarise(key_time = median(key_time),
            wpm = 60000 / (key_time * 5)) 

mturk_typing_speed$wpm %>%
  boxplot(horizontal = TRUE, main = "Typing speed of MTurk participants", xlab = "Raw words per minute")
```


## Comparison

### Fact alpha
We can compare the final alpha distributions from both data sets to see how similar the difficulty of the facts is between populations.

The final alpha values in the lab sample have a grand mean of `r mean(final_alphas$final_alpha)`.
Those in the MTurk sample have a grand mean of `r mean(final_alphas_m$final_alpha)`.
The two samples are visually quite similar, but in both cases the alpha distribution is clearly centered well above 0.3.

```{r}
final_alphas_both <- bind_rows(mutate(final_alphas, dataset = "lab"),
                               mutate(final_alphas_m, dataset = "MTurk"))

ggplot(final_alphas_both, aes(x = reorder(as.factor(fact_id), final_alpha, FUN = median), y = final_alpha, fill = dataset)) +
  geom_boxplot() +
  geom_point(alpha = 0.5, size = 0.5, position = position_jitterdodge(jitter.width = 0, dodge.width = 0.75)) +
  labs(title = "Final alpha values by fact in both samples",
       subtitle = "Only includes observations with at least 3 repetitions",
       x = "Fact",
       y = "Final alpha", 
       fill = "Sample")
```

### Typing speed

```{r}
typing_speed <- bind_rows(mutate(lab_typing_speed, dataset = "Lab"),
          mutate(mturk_typing_speed, dataset = "MTurk"))

ggplot(typing_speed, aes(y = wpm, x = dataset)) +
  geom_violin() +
  geom_jitter(width = 0.2, height = 0, size = 0.5, aes(colour = dataset)) +
  scale_y_continuous(limits = c(20, 140), breaks = seq(20, 140, by = 20)) +
  scale_colour_brewer(type = "qual", palette = 7) +
  guides(colour = FALSE) +
  labs(x = NULL,
       y = "Words per minute") +
  theme(panel.grid.minor.x = element_blank())
```

### Session stats


```{r}
session_stats <- bind_rows(mutate(session_stats, dataset = "Lab"),
                           mutate(session_stats_m, dataset = "MTurk"))
```

Number of trials:
```{r}
ggplot(session_stats, aes(y = n_trials, x = dataset)) +
  geom_violin() +
  geom_jitter(width = 0.2, height = 0, size = 0.5, aes(colour = dataset)) +
  scale_y_continuous(limits = c(50, 155)) +
  scale_colour_brewer(type = "qual", palette = 7) +
  guides(colour = FALSE) +
  labs(x = NULL,
       y = "Number of trials",
       caption = "(a)") +
  theme(panel.grid.minor.x = element_blank(),
        plot.caption = element_text(hjust = 0.5))
```

Number of facts:
```{r}
ggplot(session_stats, aes(y = n_items, x = dataset)) +
  geom_violin() +
  geom_jitter(width = 0.2, height = 0, size = 0.5, aes(colour = dataset)) +
  scale_y_continuous(limits = c(0, 30)) +
  scale_colour_brewer(type = "qual", palette = 7) +
  guides(colour = FALSE) +
  labs(x = NULL,
       y = "Number of unique facts",
       caption = "(b)") +
  theme(panel.grid.minor.x = element_blank(),
        plot.caption = element_text(hjust = 0.5))
```

Accuracy:
```{r}
ggplot(session_stats, aes(y = perc_correct, x = dataset)) +
  geom_violin() +
  geom_jitter(width = 0.2, height = 0, size = 0.5, aes(colour = dataset)) +
  scale_y_continuous(limits = c(0, 1), labels = percent_format()) +
  scale_colour_brewer(type = "qual", palette = 7) +
  guides(colour = FALSE) +
  labs(x = NULL,
       y = "Accuracy",
       caption = "(c)") +
  theme(panel.grid.minor.x = element_blank(),
        plot.caption = element_text(hjust = 0.5))
```


# Session info
```{r}
sessionInfo()
```
